Live logging
Adds "--live" option to the imu app. This outputs the json to stdout
as well as the log. The actual printing is handled by the logger.
The constructor for the logger now includes an optional "live" argument,
when this is true the logger thread prints when the data is queued.